backend/requirements.txt
fastapi
uvicorn
pydantic
kiteconnect
python-dotenv